third_party/blink/web_tests/wpt_internal/html/rendering/non-replaced-elements/flow-content-3/dialog-top-layer.html

<!doctype html>
<title>dialog: top-layer</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<dialog id="dialog"></dialog>
<script>
  const dialog = document.getElementById("dialog");

  test(() => {
    assert_equals(getComputedStyle(dialog).topLayer, "none",
                  "Computed top-layer");
    // dialog::backdrop {} UA rule always sets top-layer to 'browser' even if
    // ::backdrop pseudo is not generated. ::backdrop is only generated when the
    // dialog is in the top layer.
    assert_equals(getComputedStyle(dialog, "::backdrop").topLayer, "browser",
                  "Computed top-layer for ::backdrop");
  }, "dialog computed top-layer initially 'none'");

  test(() => {
    dialog.showModal();

    assert_equals(getComputedStyle(dialog).topLayer, "browser",
                  "Computed top-layer on open dialog");
    assert_equals(getComputedStyle(dialog, "::backdrop").topLayer, "browser",
                  "Computed top-layer for ::backdrop");

    dialog.close();

    assert_equals(getComputedStyle(dialog).topLayer, "none",
                  "Computed top-layer on closed dialog");
    // ::backdrop is always top-layer:browser. See comment above.
    assert_equals(getComputedStyle(dialog, "::backdrop").topLayer, "browser",
                  "Computed top-layer for ::backdrop");
  }, "Opening and closing a modal dialog changes computed top-layer to 'browser' and back to 'none'");
</script>
